Create utility function for updating cache with new file version
- Create `updateCachedVersions` function file in `components/files`
- Replace inline cache-updating code with utility function in
  `useUploadProjectFiles`
